Added a Basket API for React implementation

+++ backend/src/Controller/BasketController.php
@@ -4,18 +4,15 @@ namespace App\Controller;
 
 use App\Entity\Basket;
 use App\Entity\Product;
-use App\Entity\User;
 use App\Service\BasketService;
-use App\Form\BasketType;
 use App\Repository\BasketRepository;
-use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
 
 #[Route(path: '/basket')]
-
 final class BasketController extends AbstractController
 {
     private BasketService $basketService;
@@ -27,53 +24,83 @@ final class BasketController extends AbstractController
         $this->basketRepository = $basketRepository;
     }
 
-    #[Route('/',name: 'basket_view')]
-    public function viewBasket(): Response
+    #[Route('/',name: 'basket_view', methods: ['GET'])]
+    public function viewBasket(): JsonResponse
     {
         $user = $this->getUser();
+
+        if (!$user) {
+            return new JsonResponse(['error' => 'User not authenticated.'], Response::HTTP_UNAUTHORIZED);
+        }
+
         $basket = $this->basketService->getOrCreateBasket($user);
 
-        return $this->render('basket/index.html.twig', [
-            'basket' => $basket,
-        ]);
+        return new JsonResponse([
+            'basket' => $this->formatBasket($basket),
+        ], Response::HTTP_OK);
     }
 
     #[Route('/add/{id}', name: 'basket_add_product', methods: ['POST'])]
-    public function addProduct(Request $request, Product $product): Response
+    public function addProduct(Request $request, Product $product): JsonResponse
     {
-        $quantity = $request->request->get('quantity', 1);
         $user = $this->getUser();
+
+        if (!$user) {
+            return new JsonResponse(['error' => 'User not authenticated.'], Response::HTTP_UNAUTHORIZED);
+        }
+
+        $quantity = $request->request->get('quantity', 1);
         $basket = $this->basketService->getOrCreateBasket($user);
 
         $this->basketService->addProductToBasket($basket, $product, $quantity);
 
-        return $this->redirectToRoute('homepage');
+        return new JsonResponse([
+            'message' => 'Product added to basket',
+            'basket' => $this->formatBasket($basket),
+        ], Response::HTTP_CREATED);
     }
 
-    #[Route('/remove/{id}', name: 'basket_remove_product', methods: ['POST'])]
-    public function removeProduct(Product $product): Response
+    #[Route('/edit/{id}', name: 'basket_edit_product', methods: ['PUT'])]
+    public function edit(Request $request, Product $product): JsonResponse
     {
         $user = $this->getUser();
+
+        if (!$user) {
+            return new JsonResponse(['error' => 'User not authenticated.'], Response::HTTP_UNAUTHORIZED);
+        }
+
+        $quantity = $request->request->get('quantity', 1);
+        $basket = $this->basketService->getOrCreateBasket($user);
+
+        $this->basketService->updateProductQuantity($basket, $product, $quantity);
+
+        return new JsonResponse([
+            'message' => 'Product quantity edited from basket',
+            'basket' => $this->formatBasket($basket),
+        ], Response::HTTP_OK);
+    }
+
+    #[Route('/{id}', name: 'basket_remove_product', methods: ['DELETE'])]
+    public function removeProduct(Product $product): JsonResponse
+    {
+        $user = $this->getUser();
+
+        if (!$user) {
+            return new JsonResponse(['error' => 'User not authenticated.'], Response::HTTP_UNAUTHORIZED);
+        }
+
         $basket = $this->basketService->getOrCreateBasket($user);
 
         $this->basketService->removeProductFromBasket($basket, $product);
-        return $this->redirectToRoute('basket_view');
+
+        return new JsonResponse([
+            'message' => 'Product removed from basket',
+            'basket' => $this->formatBasket($basket),
+        ], Response::HTTP_OK);
     }
 
-    #[Route('/edit/{id}', name: 'basket_edit_product', methods: ['POST'])]
-    public function edit(Request $request, Product $product): Response
-    {
-        $quantity = $request->request->get('quantity');
-        $user = $this->getUser();
-        $basket = $this->basketService->getOrCreateBasket($user);
-
-        $basket = $this->basketService->updateProductQuantity($basket, $product, $quantity);
-
-        return $this->redirectToRoute('basket_view');
-    }
-
-    #[Route('clear/{id}', name: 'basket_clear', methods: ['POST'])]
-    public function clearBasket(int $id, BasketService $basketService): Response
+    #[Route('/clear/{id}', name: 'basket_clear', methods: ['POST'])]
+    public function clearBasket(int $id, BasketService $basketService): JsonResponse
     {
         $basket = $this->basketRepository->find($id);
 
@@ -83,8 +110,25 @@ final class BasketController extends AbstractController
 
         $basketService->clearBasket($basket);
 
-        return $this->redirectToRoute('basket_view', [
-            'id' => $id,
-        ]);
+        return new JsonResponse([
+            'message' => 'Basket cleared',
+        ], Response::HTTP_OK);
     }
+
+    private function formatBasket(Basket $basket): array
+    {
+        $formatted = [];
+        foreach ($basket->getBasketProducts() as $basketProduct) {
+            $formatted[] = [
+                'product' => [
+                    'id' => $basketProduct->getProduct()->getId(),
+                    'name' => $basketProduct->getProduct()->getName(),
+                    'price' => $basketProduct->getProduct()->getPrice(),
+                ],
+                'quantity' => $basketProduct->getQuantity(),
+            ];
+        }
+        return $formatted;
+    }
+
 }
